// src/app/shared/models/user.ts
export interface User {
  firstName: string;
  lastName: string;
  middleName?: string;
  birthdate: string;
  listOfInterest: string[];
  phoneNumber: string;
  house: string;
  building: string;
  barangay: string;
  street: string;
  city: string;
  province: string;
  zipcode: string;
  email: string;
  username: string;
  password: string;
  role: string;
  status: boolean;
}